refactor: move test_format into meta/header.rs
Relocate the format/parse round-trip unit test next to the Qcow2Header
code it exercises. meta/mod.rs is now purely module declarations and
re-exports. No logic changes.
